Moving Image{Push,Pull}Privileged to trust.go
