+++ cli/usage.go
@@ -16,7 +16,6 @@ var DockerCommandUsage = []Command{
 	{"login", "Log in to a Docker registry"},
 	{"logout", "Log out from a Docker registry"},
 	{"ps", "List containers"},
-	{"push", "Push an image or a repository to a registry"},
 	{"update", "Update configuration of one or more containers"},
 }
